feat: fix naming issues and improve documentation.
- Rename `filtreable` to `filterable`.
- Rename `pagineable` to `pageable`.
- Add YARD documentation to document options at the source code level.
